refactor: handle unpaid vouchers via hook
Instead giving Bank Transaction the ability to handle unpaid vouchers, we convert unpaid to paid vouchers before passing them to Bank Transaction. This feature is also extensible by other apps via hooks.
