+++ banking/hooks.py
@@ -208,6 +208,7 @@ bank_reconciliation_doctypes = [
 	"Bank Transaction",
 ]
 get_matching_queries = "banking.klarna_kosma_integration.doctype.bank_reconciliation_tool_beta.bank_reconciliation_tool_beta.get_matching_queries"
+get_payment_entries = "banking.klarna_kosma_integration.doctype.bank_reconciliation_tool_beta.unpaid_vouchers.get_payment_entries"
 
 alyf_banking_custom_records = [
 	{
